+++ src/main/java/io/coodoo/framework/listing/control/ListingQuery.java
@@ -624,6 +624,9 @@ public class ListingQuery<T> {
         if (limit != null && limit > 0) {
             typedQuery.setMaxResults(limit);
         }
+        if (ListingConfig.ORG_HIBERNATE_FETCHSIZE != 0) {
+            typedQuery.setHint("org.hibernate.fetchSize", ListingConfig.ORG_HIBERNATE_FETCHSIZE);
+        }
         return typedQuery.getResultList();
     }